Split LoginController into AuthController + PlayersController (#2057)
LoginController mixed four concerns (anonymous auth, character
selection/creation/switch, session/status reads, device telemetry) and had
the SwitchPlayer credit-then-select orchestration living directly in the
controller. Splits it into AuthController (Login/Refresh/CreateAccount/
Logout plus Status/ActiveSession/DeviceInfo) and PlayersController
(SelectPlayer/SwitchPlayer/CreatePlayer/Players/CharacterCreationData),
and extracts the SwitchPlayer/SelectPlayer orchestration into a new
Game.Api.Services.CharacterSelectionService (with PlayerDataAssembler
for the shared player-payload projection) so both controllers are back to
request validation and response mapping.

CharacterSelectionService lives in the API layer rather than folded into
AccountService, since it depends on request-scoped presentation-tier state
(SessionService, SocketManagerService) the application layer must not
reference.

Renaming the controllers changes their routes (Login/* -> Auth/* and
Players/*), so this also regenerates the frontend API client, updates the
frontend's hand-written call sites/tests, and updates every backend
integration test and doc reference to the old routes.

Closes #2057
