+++ Game.Api.Tests/Integration/PlayersControllerTests.cs
@@ -2,21 +2,15 @@ using Game.Abstractions.Contracts.Identity;
 using Game.Abstractions.DataAccess;
 using Game.Abstractions.Infrastructure;
 using Game.Api;
-using Game.Api.Controllers;
-using Game.Api.Http;
 using Game.Api.Models.Auth;
 using Game.Api.Models.Common;
 using Game.Api.Services;
 using Game.Application.Services;
 using Game.Core;
 using Game.Core.Battle.Offline;
-using Game.Core.Identity;
 using Game.Infrastructure.Database;
-using Game.Infrastructure.Entities;
 using Game.TestInfrastructure.Fixtures;
 using Game.TestInfrastructure.Helpers;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -28,126 +22,9 @@ using Xunit;
 namespace Game.Api.Tests.Integration
 {
     [Collection("Integration")]
-    public class LoginControllerTests : ApiIntegrationTestBase
+    public class PlayersControllerTests : ApiIntegrationTestBase
     {
-        public LoginControllerTests(IntegrationTestContainers containers, ITestOutputHelper testOutputHelper) : base(containers, testOutputHelper) { }
-
-        [Fact]
-        public async Task Login_ValidCredentials_ReturnsPlayerSummariesAndTokens()
-        {
-            // Arrange
-            var (_, playerId) = await SeedAsync("loginuser", "loginpass");
-
-            var creds = new { Username = "loginuser", Password = "loginpass" };
-
-            // Act
-            var response = await Client.PostAsJsonAsync("/api/Login", creds, CancellationToken);
-
-            // Assert
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse<LoginResult>>(CancellationToken);
-            Assert.NotNull(result);
-            Assert.Null(result.ErrorMessage);
-            Assert.NotNull(result.Data);
-            // Login lists the account's characters (no player is bound until SelectPlayer).
-            var summary = Assert.Single(result.Data.PlayerSummaries);
-            Assert.Equal(playerId, summary.Id);
-            Assert.Equal("TestPlayer", summary.Name);
-
-            // Both tokens are issued in the response body (no auth cookie).
-            Assert.False(response.Headers.Contains("Set-Cookie"));
-            Assert.False(string.IsNullOrEmpty(result.Data.Tokens.AccessToken));
-            Assert.False(string.IsNullOrEmpty(result.Data.Tokens.RefreshToken));
-        }
-
-        [Fact]
-        public async Task Login_IssuedAccessToken_AuthenticatesProtectedEndpoint()
-        {
-            // Arrange
-            await SeedAsync("beareruser", "bearerpass");
-
-            var (authClient, _) = await LoginAndBuildClientAsync("beareruser", "bearerpass");
-
-            // Act — the bearer access token authenticates a protected endpoint.
-            var response = await authClient.GetAsync("/api/Login/Status", CancellationToken);
-
-            // Assert
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse<Models.Player.PlayerData>>(CancellationToken);
-            Assert.NotNull(result?.Data);
-            Assert.Equal("TestPlayer", result.Data.Name);
-            authClient.Dispose();
-        }
-
-        [Fact]
-        public async Task Login_InvalidUsername_ReturnsError()
-        {
-            var creds = new { Username = "nonexistent", Password = "whatever" };
-
-            var response = await Client.PostAsJsonAsync("/api/Login", creds, CancellationToken);
-
-            Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
-
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse<LoginResult>>(CancellationToken);
-            Assert.NotNull(result);
-            Assert.NotNull(result.ErrorMessage);
-            Assert.Null(result.Data);
-        }
-
-        [Fact]
-        public async Task Login_WrongPassword_ReturnsError()
-        {
-            // Arrange — a real user whose stored hash won't match the supplied password.
-            await SeedAsync("wrongpassuser", "correctpass");
-
-            var creds = new { Username = "wrongpassuser", Password = "wrongpass" };
-
-            // Act
-            var response = await Client.PostAsJsonAsync("/api/Login", creds, CancellationToken);
-
-            // Assert — authentication is rejected and no tokens are issued.
-            Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
-
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse<LoginResult>>(CancellationToken);
-            Assert.NotNull(result);
-            Assert.NotNull(result.ErrorMessage);
-            Assert.Null(result.Data);
-        }
-
-        [Fact]
-        public async Task Login_BannedUser_ReturnsErrorAndEstablishesNoSession()
-        {
-            // A banned account with otherwise-correct credentials is rejected in the auth path: no tokens
-            // are issued and no session is established.
-            var (userId, _) = await SeedAsync("bannedlogin", "bannedpass");
-
-            using (var scope = CreateScope())
-            {
-                var context = scope.ServiceProvider.GetRequiredService<GameContext>();
-                var user = await context.Users.FindAsync([userId], CancellationToken);
-                Assert.NotNull(user);
-                user.BannedAt = DateTime.UtcNow;
-                await context.SaveChangesAsync(CancellationToken);
-            }
-
-            var creds = new { Username = "bannedlogin", Password = "bannedpass" };
-
-            // Act
-            var response = await Client.PostAsJsonAsync("/api/Login", creds, CancellationToken);
-
-            // Assert — the login is rejected with a structured error and no body data.
-            Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse<LoginResult>>(CancellationToken);
-            Assert.NotNull(result);
-            Assert.NotNull(result.ErrorMessage);
-            Assert.Null(result.Data);
-
-            // No session is established for the rejected login.
-            using var verifyScope = CreateScope();
-            var sessionStore = verifyScope.ServiceProvider.GetRequiredService<ISessionStore>();
-            Assert.Null(await sessionStore.GetSession(userId));
-        }
+        public PlayersControllerTests(IntegrationTestContainers containers, ITestOutputHelper testOutputHelper) : base(containers, testOutputHelper) { }
 
         [Fact]
         public async Task SelectPlayer_OwnedCharacter_BindsSessionAndRotatesTokenIntoGameReadyState()
@@ -173,7 +50,7 @@ namespace Game.Api.Tests.Integration
             // The rotated access token authenticates Status and resolves the selected player from its claim.
             using var authClient = Factory.CreateClient();
             authClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", select.Tokens.AccessToken);
-            var statusResponse = await authClient.GetAsync("/api/Login/Status", CancellationToken);
+            var statusResponse = await authClient.GetAsync("/api/Auth/Status", CancellationToken);
             Assert.Equal(HttpStatusCode.OK, statusResponse.StatusCode);
             var status = await statusResponse.Content.ReadFromJsonAsync<ApiResponse<Models.Player.PlayerData>>(CancellationToken);
             Assert.Equal(playerId, status?.Data?.Id);
@@ -219,11 +96,11 @@ namespace Game.Api.Tests.Integration
         [Fact]
         public async Task SelectPlayer_DeliversClassLockedBaseAndSignaturePassive()
         {
-            // The logged-in player payload (LoginController.BuildPlayerData) projects the character's class
-            // into LockedBaseDistribution + SignaturePassive — the parity-sensitive class data the live
-            // frontend battler composes its attributes from (#1126 areas D/E). The end-to-end class-delivery
-            // assertions otherwise live only on CharacterCreationData, not the logged-in payload, so a
-            // regression in this projection would load fine while desyncing the FE/BE anti-cheat surface.
+            // The logged-in player payload (PlayerDataAssembler.Build) projects the character's class into
+            // LockedBaseDistribution + SignaturePassive — the parity-sensitive class data the live frontend
+            // battler composes its attributes from (#1126 areas D/E). The end-to-end class-delivery assertions
+            // otherwise live only on CharacterCreationData, not the logged-in payload, so a regression in this
+            // projection would load fine while desyncing the FE/BE anti-cheat surface.
             using var scope = CreateScope();
             var context = scope.ServiceProvider.GetRequiredService<GameContext>();
             var user = await TestDataSeeder.CreateUserAsync(context, "classdelivery", "classpass");
@@ -300,7 +177,7 @@ namespace Game.Api.Tests.Integration
             var login = await LoginAsync("attackeracct", "pass");
 
             // The attacker selects a player they do not own — rejected before any binding.
-            using var request = new HttpRequestMessage(HttpMethod.Post, "/api/Login/SelectPlayer")
+            using var request = new HttpRequestMessage(HttpMethod.Post, "/api/Players/SelectPlayer")
             {
                 Content = JsonContent.Create(new { PlayerId = ownerPlayer.Id, login.Tokens.RefreshToken }),
             };
@@ -316,7 +193,7 @@ namespace Game.Api.Tests.Integration
         [Fact]
         public async Task SelectPlayer_Unauthenticated_Returns401()
         {
-            var response = await Client.PostAsJsonAsync("/api/Login/SelectPlayer",
+            var response = await Client.PostAsJsonAsync("/api/Players/SelectPlayer",
                 new { PlayerId = 1, RefreshToken = "irrelevant" }, CancellationToken);
 
             Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
@@ -325,7 +202,7 @@ namespace Game.Api.Tests.Integration
         [Fact]
         public async Task SwitchPlayer_Unauthenticated_Returns401()
         {
-            var response = await Client.PostAsJsonAsync("/api/Login/SwitchPlayer",
+            var response = await Client.PostAsJsonAsync("/api/Players/SwitchPlayer",
                 new { PlayerId = 1, RefreshToken = "irrelevant" }, CancellationToken);
 
             Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
@@ -348,7 +225,7 @@ namespace Game.Api.Tests.Integration
             var select = await SelectPlayerAsync(login.Tokens, attackerPlayer.Id);
 
             // Switching to a character of another account is rejected (anti-cheat) before binding it.
-            using var request = new HttpRequestMessage(HttpMethod.Post, "/api/Login/SwitchPlayer")
+            using var request = new HttpRequestMessage(HttpMethod.Post, "/api/Players/SwitchPlayer")
             {
                 Content = JsonContent.Create(new { PlayerId = ownerPlayer.Id, select.Tokens.RefreshToken }),
             };
@@ -415,7 +292,7 @@ namespace Game.Api.Tests.Integration
             // The rotated access token authenticates Status and resolves the target character from its claim.
             using var authClient = Factory.CreateClient();
             authClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", switched.Tokens.AccessToken);
-            var statusResponse = await authClient.GetAsync("/api/Login/Status", CancellationToken);
+            var statusResponse = await authClient.GetAsync("/api/Auth/Status", CancellationToken);
             Assert.Equal(HttpStatusCode.OK, statusResponse.StatusCode);
             var status = await statusResponse.Content.ReadFromJsonAsync<ApiResponse<Models.Player.PlayerData>>(CancellationToken);
             Assert.Equal(target.Id, status?.Data?.Id);
@@ -541,10 +418,10 @@ namespace Game.Api.Tests.Integration
             await context.SaveChangesAsync(CancellationToken);
             await ReloadReferenceCachesAsync();
 
-            // A LoginController built by hand, swapping in an OfflineProgressService whose progress save always
-            // faults — a targeted stand-in for a genuine SimulateSwitchProgress failure (a transient Redis/DB
-            // blip) that leaves the claim-release finally block as the only thing standing between this and a
-            // stuck claim.
+            // A CharacterSelectionService built by hand, swapping in an OfflineProgressService whose progress
+            // save always faults — a targeted stand-in for a genuine SimulateSwitchProgress failure (a
+            // transient Redis/DB blip) that leaves the claim-release finally block as the only thing standing
+            // between this and a stuck claim.
             var faultingOfflineService = new OfflineProgressService(
                 scope.ServiceProvider.GetRequiredService<IPlayerRepository>(),
                 new SaveThrowsPlayerProgressRepository(scope.ServiceProvider.GetRequiredService<IPlayerProgressRepository>()),
@@ -563,23 +440,22 @@ namespace Game.Api.Tests.Integration
             var faultingSessionService = new SessionService(scope.ServiceProvider.GetRequiredService<ISessionStore>());
             faultingSessionService.SetAuthenticatedUser(user.Id, departed.Id);
 
-            var controller = new LoginController(
+            var playerDataAssembler = new PlayerDataAssembler(
+                scope.ServiceProvider.GetRequiredService<IClasses>(),
+                scope.ServiceProvider.GetRequiredService<BattleService>());
+
+            var characterSelectionService = new CharacterSelectionService(
                 faultingSessionService,
                 scope.ServiceProvider.GetRequiredService<SessionInitializer>(),
                 scope.ServiceProvider.GetRequiredService<AccountService>(),
-                scope.ServiceProvider.GetRequiredService<LoginTrackingService>(),
                 scope.ServiceProvider.GetRequiredService<SocketManagerService>(),
                 scope.ServiceProvider.GetRequiredService<PlayerService>(),
                 faultingOfflineService,
-                scope.ServiceProvider.GetRequiredService<BattleService>(),
-                scope.ServiceProvider.GetRequiredService<IClasses>(),
-                scope.ServiceProvider.GetRequiredService<ILogger<LoginController>>())
-            {
-                ControllerContext = new ControllerContext { HttpContext = new DefaultHttpContext() },
-            };
+                playerDataAssembler,
+                scope.ServiceProvider.GetRequiredService<ILogger<CharacterSelectionService>>());
 
             await Assert.ThrowsAsync<InvalidOperationException>(
-                () => controller.SwitchPlayer(new SelectPlayerRequest { PlayerId = target.Id, RefreshToken = "unused" }));
+                () => characterSelectionService.SwitchPlayer(user.Id, target.Id, "unused", CancellationToken));
 
             var cache = scope.ServiceProvider.GetRequiredService<ICacheService>();
             var presenceKey = $"{Constants.CACHE_PLAYER_SOCKET_PREFIX}_{departed.Id}";
@@ -617,7 +493,7 @@ namespace Game.Api.Tests.Integration
 
             using var authClient = Factory.CreateClient();
             authClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", select.Tokens.AccessToken);
-            var response = await authClient.GetAsync("/api/Login/Players", CancellationToken);
+            var response = await authClient.GetAsync("/api/Players", CancellationToken);
 
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
             var result = await response.Content.ReadFromJsonAsync<ApiEnumerableResponse<PlayerSummary>>(CancellationToken);
@@ -628,7 +504,7 @@ namespace Game.Api.Tests.Integration
         [Fact]
         public async Task Players_Unauthenticated_Returns401()
         {
-            var response = await Client.GetAsync("/api/Login/Players", CancellationToken);
+            var response = await Client.GetAsync("/api/Players", CancellationToken);
             Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
         }
 
@@ -636,7 +512,7 @@ namespace Game.Api.Tests.Integration
         // result (rotated tokens plus the loaded target player).
         private async Task<SelectPlayerResult> SwitchPlayerAsync(AuthTokens tokens, int playerId)
         {
-            using var request = new HttpRequestMessage(HttpMethod.Post, "/api/Login/SwitchPlayer")
+            using var request = new HttpRequestMessage(HttpMethod.Post, "/api/Players/SwitchPlayer")
             {
                 Content = JsonContent.Create(new { PlayerId = playerId, tokens.RefreshToken }),
             };
@@ -673,91 +549,6 @@ namespace Game.Api.Tests.Integration
         }
 
         [Fact]
-        public async Task CreateAccount_ValidCredentials_Succeeds()
-        {
-            // Signup creates the account only — no character, so no class is supplied (#1256).
-            var creds = new { Username = "newuser", Password = "newpass" };
-
-            var response = await Client.PostAsJsonAsync("/api/Login/CreateAccount", creds, CancellationToken);
-
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse>(CancellationToken);
-            Assert.NotNull(result);
-            Assert.Null(result.ErrorMessage);
-        }
-
-        [Fact]
-        public async Task CreateAccount_DuplicateUsername_ReturnsError()
-        {
-            // Arrange — create the user first so the duplicate-username check rejects the second attempt.
-            using (var scope = CreateScope())
-            {
-                var context = scope.ServiceProvider.GetRequiredService<GameContext>();
-                await TestDataSeeder.CreateUserAsync(context, "duplicate", "pass");
-            }
-
-            var creds = new { Username = "duplicate", Password = "anotherpass" };
-
-            // Act
-            var response = await Client.PostAsJsonAsync("/api/Login/CreateAccount", creds, CancellationToken);
-
-            // Assert
-            Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse>(CancellationToken);
-            Assert.NotNull(result);
-            Assert.NotNull(result.ErrorMessage);
-        }
-
-        [Fact]
-        public async Task CreateAccount_UsernameAtMaxLength_Succeeds()
-        {
-            // Exactly the 20-char column limit (UsernamePolicy.MaxLength) — the boundary that must still fit storage.
-            var creds = new { Username = new string('u', UsernamePolicy.MaxLength), Password = "newpass" };
-
-            var response = await Client.PostAsJsonAsync("/api/Login/CreateAccount", creds, CancellationToken);
-
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse>(CancellationToken);
-            Assert.NotNull(result);
-            Assert.Null(result.ErrorMessage);
-        }
-
-        [Fact]
-        public async Task CreateAccount_UsernameOverMaxLength_ReturnsValidationErrorNotServerError()
-        {
-            // One character past the 20-char column limit — [ApiController]'s automatic model validation must
-            // reject it (as a ValidationProblemDetails 400, its standard shape for a DataAnnotations failure)
-            // before it ever reaches the database, rather than letting a Postgres string-truncation error
-            // escape the self-commit as a 500.
-            var creds = new { Username = new string('u', UsernamePolicy.MaxLength + 1), Password = "newpass" };
-
-            var response = await Client.PostAsJsonAsync("/api/Login/CreateAccount", creds, CancellationToken);
-
-            Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
-            var problem = await response.Content.ReadFromJsonAsync<ValidationProblemDetails>(CancellationToken);
-            Assert.NotNull(problem);
-            Assert.Contains(nameof(CreateAccountRequest.Username), problem.Errors.Keys);
-        }
-
-        [Fact]
-        public async Task CreateAccount_ConcurrentDuplicate_ReturnsCleanErrorNotServerError()
-        {
-            var creds = new { Username = "raceuser", Password = "racepass" };
-
-            // Two concurrent requests racing past the existence check both reach the commit. The
-            // active-username unique index lets exactly one through; the loser must surface as a clean
-            // BadRequest, not the 500 the violation would otherwise raise outside the action.
-            var responses = await Task.WhenAll(
-                Client.PostAsJsonAsync("/api/Login/CreateAccount", creds, CancellationToken),
-                Client.PostAsJsonAsync("/api/Login/CreateAccount", creds, CancellationToken));
-
-            Assert.Equal(1, responses.Count(response => response.StatusCode == HttpStatusCode.OK));
-            Assert.Equal(1, responses.Count(response => response.StatusCode == HttpStatusCode.BadRequest));
-            Assert.DoesNotContain(responses, response => response.StatusCode == HttpStatusCode.InternalServerError);
-        }
-
-        [Fact]
         public async Task CreatePlayer_AuthenticatedValidName_CreatesCharacterAndReturnsSummary()
         {
             // The class kit (its starter skills) backs the new-player blueprint's PlayerSkills.
@@ -777,7 +568,7 @@ namespace Game.Api.Tests.Integration
             // A pre-selection token (no player bound) — the realistic state on the character-select screen.
             TestAuthHelper.AddAuthHeader(client, user.Id);
 
-            var response = await client.PostAsJsonAsync("/api/Login/CreatePlayer", new { Name = "Gandalf", ClassId = classId }, CancellationToken);
+            var response = await client.PostAsJsonAsync("/api/Players/CreatePlayer", new { Name = "Gandalf", ClassId = classId }, CancellationToken);
 
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
             var result = await response.Content.ReadFromJsonAsync<ApiResponse<PlayerSummary>>(CancellationToken);
@@ -798,7 +589,7 @@ namespace Game.Api.Tests.Integration
         [Fact]
         public async Task CreatePlayer_Unauthenticated_Returns401()
         {
-            var response = await Client.PostAsJsonAsync("/api/Login/CreatePlayer", new { Name = "Nobody" }, CancellationToken);
+            var response = await Client.PostAsJsonAsync("/api/Players/CreatePlayer", new { Name = "Nobody" }, CancellationToken);
 
             Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
         }
@@ -815,7 +606,7 @@ namespace Game.Api.Tests.Integration
 
             // A name past the 20-char limit is rejected with a structured error, not a 500. The name is
             // validated before the class, so the placeholder class id is never reached.
-            var response = await client.PostAsJsonAsync("/api/Login/CreatePlayer",
+            var response = await client.PostAsJsonAsync("/api/Players/CreatePlayer",
                 new { Name = "this-name-is-way-too-long-to-be-valid", ClassId = 0 }, CancellationToken);
 
             Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
@@ -848,389 +639,13 @@ namespace Game.Api.Tests.Integration
             var client = Factory.CreateClient();
             TestAuthHelper.AddAuthHeader(client, user.Id);
 
-            var response = await client.PostAsJsonAsync("/api/Login/CreatePlayer", new { Name = "OneTooMany", ClassId = classId }, CancellationToken);
+            var response = await client.PostAsJsonAsync("/api/Players/CreatePlayer", new { Name = "OneTooMany", ClassId = classId }, CancellationToken);
 
             Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
             using var verifyScope = CreateScope();
             var verifyContext = verifyScope.ServiceProvider.GetRequiredService<GameContext>();
             Assert.Equal(6, await verifyContext.Players.CountAsync(p => p.UserId == user.Id, CancellationToken));
             client.Dispose();
-        }
-
-        [Fact]
-        public async Task Status_Unauthenticated_Returns401()
-        {
-            var response = await Client.GetAsync("/api/Login/Status", CancellationToken);
-
-            Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
-        }
-
-        [Fact]
-        public async Task Status_AuthenticatedButPlayerNotLoadable_Returns404WithError()
-        {
-            // A still-valid token whose player can't be loaded (archived/deleted between requests; here a user
-            // with no player at all) must return a structured error, not a 500 — mirroring how Login surfaces a
-            // missing player. Rehydration finds no player, so the session's selected id stays unresolved and the
-            // player load comes back null. The missing-resource semantics surface as 404, not a blanket 400.
-            using var scope = CreateScope();
-            var context = scope.ServiceProvider.GetRequiredService<GameContext>();
-            var user = await TestDataSeeder.CreateUserAsync(context, "playerlessstatus", "pass");
-
-            var client = Factory.CreateClient();
-            TestAuthHelper.AddAuthHeader(client, user.Id);
-
-            var response = await client.GetAsync("/api/Login/Status", CancellationToken);
-
-            Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse<Models.Player.PlayerData>>(CancellationToken);
-            Assert.Null(result?.Data);
-            Assert.NotNull(result?.ErrorMessage);
-            client.Dispose();
-        }
-
-        [Fact]
-        public async Task ActiveSession_Unauthenticated_Returns401()
-        {
-            var response = await Client.GetAsync("/api/Login/ActiveSession", CancellationToken);
-
-            Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
-        }
-
-        [Fact]
-        public async Task Status_ValidTokenWithNoSessionCache_RehydratesAndReturnsPlayer()
-        {
-            // A valid token with no cached session (evicted, aged out under the sliding TTL, or never
-            // established on this instance) must not be reported as "not logged in" (#693). The session is
-            // rehydrated from the user's player binding instead.
-            var (client, userId, _) = await SeedUserWithTokenButNoSessionAsync("evictedstatus");
-
-            var response = await client.GetAsync("/api/Login/Status", CancellationToken);
-
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse<Models.Player.PlayerData>>(CancellationToken);
-            Assert.NotNull(result?.Data);
-            Assert.Null(result.ErrorMessage);
-            Assert.Equal("TestPlayer", result.Data.Name);
-
-            // Rehydration is in-memory only: the request resolves the player without ever writing the session
-            // cache, since player-state writes belong on the socket, not this concurrent HTTP path (#937).
-            await AssertSessionNotEstablishedAsync(userId);
-            client.Dispose();
-        }
-
-        [Fact]
-        public async Task ActiveSession_ValidTokenWithNoSessionCache_RehydratesAndReturnsResult()
-        {
-            // The pre-game active-session takeover warning is the user-visible breakage from #693: an evicted
-            // session must rehydrate and report the (absent) active socket, not "not logged in".
-            var (client, userId, _) = await SeedUserWithTokenButNoSessionAsync("evictedactive");
-
-            var response = await client.GetAsync("/api/Login/ActiveSession", CancellationToken);
-
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse<ActiveSessionResult>>(CancellationToken);
-            Assert.NotNull(result?.Data);
-            Assert.Null(result.ErrorMessage);
-            Assert.False(result.Data.Active);
-
-            // Rehydration is in-memory only — the presence check resolves the player without writing the cache (#937).
-            await AssertSessionNotEstablishedAsync(userId);
-            client.Dispose();
-        }
-
-        [Fact]
-        public async Task NonSessionEndpoint_ValidTokenWithNoSessionCache_DoesNotEstablishSession()
-        {
-            // The redundant per-request session read is gone (#755): an authenticated endpoint that never
-            // reads player state (here DeviceInfo) must not load or rehydrate the session cache, even for a
-            // user who has a resolvable player. Only the socket handshake and Status/ActiveSession do.
-            var (client, userId, _) = await SeedUserWithTokenButNoSessionAsync("nosessionread");
-            client.DefaultRequestHeaders.TryAddWithoutValidation(ClientHints.DeviceFingerprintHeader, "fp-nosession");
-
-            var response = await client.PostAsJsonAsync("/api/Login/DeviceInfo",
-                new { DeviceMemory = 8.0, HardwareConcurrency = 4 }, CancellationToken);
-
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-
-            // Confirm the session was never established — the cache stays empty for this user.
-            await AssertSessionNotEstablishedAsync(userId);
-            client.Dispose();
-        }
-
-        /// <summary>
-        /// Seeds a user with a player (and a linked skill so the aggregate loads) and returns a client
-        /// carrying a valid bearer token for that user but with no session ever established in the cache —
-        /// the "valid token, evicted/absent session" state.
-        /// </summary>
-        private async Task<(HttpClient Client, int UserId, int PlayerId)>
-            SeedUserWithTokenButNoSessionAsync(string username)
-        {
-            var (userId, playerId) = await SeedAsync(username, "pass");
-
-            using (var scope = CreateScope())
-            {
-                var sessionStore = scope.ServiceProvider.GetRequiredService<ISessionStore>();
-                Assert.Null(await sessionStore.GetSession(userId));
-            }
-
-            var client = Factory.CreateClient();
-            // A post-selection token (carrying the selected-player claim) with no cached session — the
-            // "valid token, evicted/absent session" state that must rehydrate from the claim.
-            TestAuthHelper.AddAuthHeader(client, userId, playerId);
-            return (client, userId, playerId);
-        }
-
-        // Confirms a session was never written to the cache: rehydration (and any non-session endpoint) resolves
-        // the player in memory only, so we give any erroneous fire-and-forget write a window to land, then assert
-        // the key stays absent for this user.
-        private async Task AssertSessionNotEstablishedAsync(int userId)
-        {
-            await Task.Delay(250, CancellationToken);
-            using var scope = CreateScope();
-            var sessionStore = scope.ServiceProvider.GetRequiredService<ISessionStore>();
-            Assert.Null(await sessionStore.GetSession(userId));
-        }
-
-        [Fact]
-        public async Task ActiveSession_NoOpenSocket_ReturnsFalse()
-        {
-            // Arrange — a logged-in user who has not opened a game connection.
-            await SeedAsync("nosocketuser", "nosocketpass");
-
-            var (authClient, _) = await LoginAndBuildClientAsync("nosocketuser", "nosocketpass");
-
-            // Act
-            var result = await GetActiveSessionAsync(authClient);
-
-            // Assert — no live connection means no other session to warn about.
-            Assert.NotNull(result?.Data);
-            Assert.False(result.Data.Active);
-            authClient.Dispose();
-        }
-
-        [Fact]
-        public async Task ActiveSession_WithOpenSocket_ReturnsTrue()
-        {
-            // Arrange — a logged-in user with a live websocket connection registered.
-            var (userId, _) = await SeedAsync("livesocketuser", "livesocketpass");
-
-            var (authClient, _) = await LoginAndBuildClientAsync("livesocketuser", "livesocketpass");
-
-            await using var socketClient = new TestSocketClient();
-            var wsClient = Factory.Server.CreateWebSocketClient();
-            await socketClient.ConnectAsync(wsClient, userId);
-            // Round-trip a command so the connection is fully registered (the socket-presence key is set
-            // before the command listener, so any response guarantees registration completed).
-            await socketClient.SendCommandAsync<object>("GetStatisticTypes");
-
-            // Act
-            var result = await GetActiveSessionAsync(authClient);
-
-            // Assert — the open connection is reported as an active session.
-            Assert.NotNull(result?.Data);
-            Assert.True(result.Data.Active);
-            authClient.Dispose();
-        }
-
-        private static async Task<ApiResponse<ActiveSessionResult>?> GetActiveSessionAsync(HttpClient client)
-        {
-            var response = await client.GetAsync("/api/Login/ActiveSession", CancellationToken);
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            return await response.Content.ReadFromJsonAsync<ApiResponse<ActiveSessionResult>>(CancellationToken);
-        }
-
-        [Fact]
-        public async Task Refresh_ValidToken_RotatesAndReturnsNewTokens()
-        {
-            // Arrange
-            var (_, playerId) = await SeedAsync("refreshuser", "refreshpass");
-
-            // Select a character so the refreshed token carries the selected player (and Status can load it).
-            var login = await LoginAsync("refreshuser", "refreshpass");
-            var select = await SelectPlayerAsync(login.Tokens, playerId);
-
-            // Act — exchange the refresh token for a new pair.
-            var refreshResponse = await Client.PostAsJsonAsync("/api/Login/Refresh",
-                new { select.Tokens.RefreshToken }, CancellationToken);
-
-            // Assert — a fresh, rotated pair is returned.
-            Assert.Equal(HttpStatusCode.OK, refreshResponse.StatusCode);
-            var refreshed = await refreshResponse.Content.ReadFromJsonAsync<ApiResponse<AuthTokens>>(CancellationToken);
-            Assert.NotNull(refreshed?.Data);
-            Assert.False(string.IsNullOrEmpty(refreshed.Data.AccessToken));
-            Assert.False(string.IsNullOrEmpty(refreshed.Data.RefreshToken));
-            Assert.NotEqual(select.Tokens.RefreshToken, refreshed.Data.RefreshToken);
-
-            // The new access token authenticates a protected endpoint and keeps the selected player bound.
-            using var authClient = Factory.CreateClient();
-            authClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", refreshed.Data.AccessToken);
-            var statusResponse = await authClient.GetAsync("/api/Login/Status", CancellationToken);
-            Assert.Equal(HttpStatusCode.OK, statusResponse.StatusCode);
-        }
-
-        [Fact]
-        public async Task Refresh_SameTokenTwice_IsRejectedSecondTime()
-        {
-            // Arrange
-            await SeedAsync("rotateuser", "rotatepass");
-
-            var login = await LoginAsync("rotateuser", "rotatepass");
-
-            // Act — first use succeeds, replaying the same (now consumed) token fails.
-            var first = await Client.PostAsJsonAsync("/api/Login/Refresh",
-                new { RefreshToken = login.Tokens.RefreshToken }, CancellationToken);
-            Assert.Equal(HttpStatusCode.OK, first.StatusCode);
-
-            var second = await Client.PostAsJsonAsync("/api/Login/Refresh",
-                new { RefreshToken = login.Tokens.RefreshToken }, CancellationToken);
-
-            // Assert — single-use rotation means the original token is no longer valid.
-            Assert.Equal(HttpStatusCode.BadRequest, second.StatusCode);
-            var result = await second.Content.ReadFromJsonAsync<ApiResponse<AuthTokens>>(CancellationToken);
-            Assert.NotNull(result);
-            Assert.NotNull(result.ErrorMessage);
-        }
-
-        [Fact]
-        public async Task Refresh_InvalidToken_ReturnsError()
-        {
-            var response = await Client.PostAsJsonAsync("/api/Login/Refresh",
-                new { RefreshToken = "not-a-real-token" }, CancellationToken);
-
-            Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse<AuthTokens>>(CancellationToken);
-            Assert.NotNull(result);
-            Assert.NotNull(result.ErrorMessage);
-        }
-
-        [Fact]
-        public async Task Logout_Authenticated_RevokesRefreshTokenAndEndsSession()
-        {
-            // Arrange — a logged-in user carrying a valid access token.
-            var (userId, _) = await SeedAsync("logoutuser", "logoutpass");
-
-            var (authClient, tokens) = await LoginAndBuildClientAsync("logoutuser", "logoutpass");
-
-            // Act
-            var response = await authClient.PostAsJsonAsync("/api/Login/Logout",
-                new { tokens.RefreshToken }, CancellationToken);
-
-            // Assert — logout succeeds, the session is cleared, and the refresh token is revoked.
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse>(CancellationToken);
-            Assert.NotNull(result);
-            Assert.Null(result.ErrorMessage);
-
-            using var scope = CreateScope();
-            var cache = scope.ServiceProvider.GetRequiredService<ICacheService>();
-            var session = await cache.Get($"Session_{userId}");
-            Assert.Null(session);
-
-            // The revoked refresh token can no longer be exchanged for new tokens.
-            var refreshResponse = await Client.PostAsJsonAsync("/api/Login/Refresh",
-                new { tokens.RefreshToken }, CancellationToken);
-            Assert.Equal(HttpStatusCode.BadRequest, refreshResponse.StatusCode);
-            authClient.Dispose();
-        }
-
-        [Fact]
-        public async Task Logout_ExpiredAccessTokenButValidRefreshToken_StillEvictsSession()
-        {
-            // The common logout path (#906): the 15-minute access token has already expired, so the client
-            // logs out anonymously with just its refresh token. No request principal means no recorded
-            // UserId, yet the cached session must still be evicted — derived from the consumed refresh token.
-            var (userId, playerId) = await SeedAsync("expiredlogout", "logoutpass");
-
-            // Selecting a character establishes the cached session; the refresh token outlives the access token.
-            var login = await LoginAsync("expiredlogout", "logoutpass");
-            var select = await SelectPlayerAsync(login.Tokens, playerId);
-            await AssertSessionPresentAsync(userId);
-
-            // Act — log out over the unauthenticated client (no bearer token, mimicking the expired access
-            // token) carrying only the still-valid refresh token.
-            var response = await Client.PostAsJsonAsync("/api/Login/Logout",
-                new { select.Tokens.RefreshToken }, CancellationToken);
-
-            // Assert — logout succeeds and the session is evicted despite the absent access token.
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            await AssertSessionEvictedAsync(userId);
-
-            // The consumed refresh token can no longer be exchanged for new tokens.
-            var refreshResponse = await Client.PostAsJsonAsync("/api/Login/Refresh",
-                new { select.Tokens.RefreshToken }, CancellationToken);
-            Assert.Equal(HttpStatusCode.BadRequest, refreshResponse.StatusCode);
-        }
-
-        [Fact]
-        public async Task Logout_Unauthenticated_Succeeds()
-        {
-            // Logout is AllowAnonymous so it always succeeds, even without a valid session/token.
-            var response = await Client.PostAsJsonAsync("/api/Login/Logout",
-                new { RefreshToken = "irrelevant" }, CancellationToken);
-
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            var result = await response.Content.ReadFromJsonAsync<ApiResponse>(CancellationToken);
-            Assert.NotNull(result);
-            Assert.Null(result.ErrorMessage);
-        }
-
-        // The session store write after login is fire-and-forget, so poll until the session is cached.
-        private async Task AssertSessionPresentAsync(int userId)
-        {
-            using var scope = CreateScope();
-            var sessionStore = scope.ServiceProvider.GetRequiredService<ISessionStore>();
-            var session = await PollingHelper.PollUntilAsync(
-                () => sessionStore.GetSession(userId), s => s is not null, timeoutMs: 5000);
-
-            Assert.True(session is not null, "The session was not established in the cache after login.");
-        }
-
-        // The Clear on logout is fire-and-forget, so poll until the session disappears.
-        private async Task AssertSessionEvictedAsync(int userId)
-        {
-            using var scope = CreateScope();
-            var sessionStore = scope.ServiceProvider.GetRequiredService<ISessionStore>();
-            var session = await PollingHelper.PollUntilAsync(
-                () => sessionStore.GetSession(userId), s => s is null, timeoutMs: 5000);
-
-            Assert.True(session is null, "The session was not evicted from the cache on logout.");
-        }
-
-        [Fact]
-        public async Task Login_AdminUser_InjectsRoleIntoTokenAndGrantsAdminAccess()
-        {
-            // Arrange — a user granted the seeded Admin role.
-            var (userId, _) = await SeedAsync("adminlogin", "adminpass");
-            using (var scope = CreateScope())
-            {
-                var context = scope.ServiceProvider.GetRequiredService<GameContext>();
-                await TestDataSeeder.AssignRoleToUserAsync(context, userId, ERole.Admin);
-            }
-
-            // Act — log in and reuse the issued access token against an admin endpoint.
-            var (authClient, _) = await LoginAndBuildClientAsync("adminlogin", "adminpass");
-
-            var response = await authClient.PostAsJsonAsync("/api/AdminTools/AddEditTags", Array.Empty<object>(), CancellationToken);
-
-            // Assert — the role baked into the token grants access (no 401/403).
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            authClient.Dispose();
-        }
-
-        [Fact]
-        public async Task Login_NonAdminUser_DoesNotGrantAdminAccess()
-        {
-            // Arrange — a user without any roles.
-            await SeedAsync("plainlogin", "plainpass");
-
-            // Act
-            var (authClient, _) = await LoginAndBuildClientAsync("plainlogin", "plainpass");
-
-            var response = await authClient.PostAsJsonAsync("/api/AdminTools/AddEditTags", Array.Empty<object>(), CancellationToken);
-
-            // Assert — authenticated, but lacking the Admin role.
-            Assert.Equal(HttpStatusCode.Forbidden, response.StatusCode);
-            authClient.Dispose();
         }
 
         [Fact]
@@ -1258,7 +673,7 @@ namespace Game.Api.Tests.Integration
             authClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", login.Tokens.AccessToken);
 
             // Act
-            var response = await authClient.GetAsync("/api/Login/CharacterCreationData", CancellationToken);
+            var response = await authClient.GetAsync("/api/Players/CharacterCreationData", CancellationToken);
 
             // Assert
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
@@ -1296,7 +711,7 @@ namespace Game.Api.Tests.Integration
         public async Task CharacterCreationData_WithoutAuthentication_IsRejected()
         {
             // No bearer token — a pre-authentication request never reaches the creatable-class payload.
-            var response = await Client.GetAsync("/api/Login/CharacterCreationData", CancellationToken);
+            var response = await Client.GetAsync("/api/Players/CharacterCreationData", CancellationToken);
             Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
         }
     }
