+++ UI/src/lib/api/types/api-type-map.ts
@@ -103,18 +103,18 @@ export type ApiResponseTypes = {
 	'AdminTools/SetZoneEnemies': undefined;
 	'AdminTools/UnarchiveUser': undefined;
 	'AdminTools/UnbanUser': undefined;
-	'Login': ILoginResult;
-	'Login/ActiveSession': IActiveSessionResult;
-	'Login/CharacterCreationData': ICreatableClass[];
-	'Login/CreateAccount': undefined;
-	'Login/CreatePlayer': IPlayerSummary;
-	'Login/DeviceInfo': undefined;
-	'Login/Logout': undefined;
-	'Login/Players': IPlayerSummary[];
-	'Login/Refresh': IAuthTokens;
-	'Login/SelectPlayer': ISelectPlayerResult;
-	'Login/Status': IPlayerData;
-	'Login/SwitchPlayer': ISelectPlayerResult;
+	'Auth': ILoginResult;
+	'Auth/ActiveSession': IActiveSessionResult;
+	'Auth/CreateAccount': undefined;
+	'Auth/DeviceInfo': undefined;
+	'Auth/Logout': undefined;
+	'Auth/Refresh': IAuthTokens;
+	'Auth/Status': IPlayerData;
+	'Players': IPlayerSummary[];
+	'Players/CharacterCreationData': ICreatableClass[];
+	'Players/CreatePlayer': IPlayerSummary;
+	'Players/SelectPlayer': ISelectPlayerResult;
+	'Players/SwitchPlayer': ISelectPlayerResult;
 	'Tags': ITag[];
 	'Tags/TagCategories': ITagCategory[];
 };
@@ -163,14 +163,14 @@ export type ApiRequestTypes = {
 	'AdminTools/SetZoneEnemies': ISetZoneEnemiesData;
 	'AdminTools/UnarchiveUser': IUserActionData;
 	'AdminTools/UnbanUser': IUserActionData;
-	'Login': ILoginCredentials;
-	'Login/CreateAccount': ICreateAccountRequest;
-	'Login/CreatePlayer': ICreatePlayerRequest;
-	'Login/DeviceInfo': IDeviceInfoRequest;
-	'Login/Logout': IRefreshRequest;
-	'Login/Refresh': IRefreshRequest;
-	'Login/SelectPlayer': ISelectPlayerRequest;
-	'Login/SwitchPlayer': ISelectPlayerRequest;
+	'Auth': ILoginCredentials;
+	'Auth/CreateAccount': ICreateAccountRequest;
+	'Auth/DeviceInfo': IDeviceInfoRequest;
+	'Auth/Logout': IRefreshRequest;
+	'Auth/Refresh': IRefreshRequest;
+	'Players/CreatePlayer': ICreatePlayerRequest;
+	'Players/SelectPlayer': ISelectPlayerRequest;
+	'Players/SwitchPlayer': ISelectPlayerRequest;
 };
 
 export type ApiEndpoint = keyof ApiResponseTypes;
