+++ docs/backend-auth.md
@@ -8,7 +8,7 @@ Passwords use **PBKDF2-HMAC-SHA256** behind an `IPasswordHasher` abstraction. A 
 
 ## Account/login orchestration
 
-Account creation and the login/select/refresh/logout flows are orchestrated by `AccountService` in the application layer; the controller is a thin HTTP adapter that maps the result and wires the request-scoped session (a presentation concern). The new-player defaults are a domain concern built by a `NewPlayerFactory` in `Game.Core`, so the application layer neither encodes the defaults nor constructs entity graphs. **Signup creates the account only — no character** (#1256): `CreateAccount` takes just credentials, so the class-selection surface is a single one (the create-character screen) rather than being duplicated into the auth form. **Class drives creation:** every character is created through `CreatePlayer` — the account's first on the select screen right after signup, and any additional ones — which takes a `classId`, resolves it against the class catalogue (rejecting an unknown or retired class), and hands the resolved `Class` to the factory, which produces the character's starter kit and starting attribute spread from class data. The starter kit also includes **equipped** starting items (a class's `StarterEquipment`, the one new-player mechanism with no prior precedent — players carried no inventory at creation before): the blueprint carries them as raw `(ItemId, EEquipmentSlot)` pairs and the data tier persists each as an unlocked-and-equipped row, so the weapon's innate `GrantedSkillId` comes online through the normal equip path at battle assembly. The slot↔category match is an authoring-time guard (`AdminClasses.SetStarterEquipment`), so creation trusts the kit as well-formed, exactly as it trusts the starter skills' `Player`-acquirable flag. The chosen class persists as the non-nullable `Player.ClassId` seam — permanent (a different archetype is a new character) and the anchor the durable class effects (locked-base fingerprint, signature passive) build on. Access-token issuance sits behind an `IAccessTokenService` abstraction so the application layer stays free of the JWT libraries, while the concrete implementation lives at the presentation edge alongside the bearer-validation pipeline.
+Account creation and the login/select/refresh/logout flows are orchestrated by `AccountService` in the application layer; the controllers (`AuthController` for the anonymous flow plus session/status reads, `PlayersController` for character selection/creation/switch, #2057) are thin HTTP adapters that map the result and wire the request-scoped session (a presentation concern). The character-selection and in-game-switch orchestration — binding the session, and (for a switch) crediting the departed character — is one step removed from the controller: `Game.Api.Services.CharacterSelectionService` owns it, since it depends on request-scoped presentation-tier state (`SessionService`, `SocketManagerService`) the application layer must not reference. The new-player defaults are a domain concern built by a `NewPlayerFactory` in `Game.Core`, so the application layer neither encodes the defaults nor constructs entity graphs. **Signup creates the account only — no character** (#1256): `CreateAccount` takes just credentials, so the class-selection surface is a single one (the create-character screen) rather than being duplicated into the auth form. **Class drives creation:** every character is created through `CreatePlayer` — the account's first on the select screen right after signup, and any additional ones — which takes a `classId`, resolves it against the class catalogue (rejecting an unknown or retired class), and hands the resolved `Class` to the factory, which produces the character's starter kit and starting attribute spread from class data. The starter kit also includes **equipped** starting items (a class's `StarterEquipment`, the one new-player mechanism with no prior precedent — players carried no inventory at creation before): the blueprint carries them as raw `(ItemId, EEquipmentSlot)` pairs and the data tier persists each as an unlocked-and-equipped row, so the weapon's innate `GrantedSkillId` comes online through the normal equip path at battle assembly. The slot↔category match is an authoring-time guard (`AdminClasses.SetStarterEquipment`), so creation trusts the kit as well-formed, exactly as it trusts the starter skills' `Player`-acquirable flag. The chosen class persists as the non-nullable `Player.ClassId` seam — permanent (a different archetype is a new character) and the anchor the durable class effects (locked-base fingerprint, signature passive) build on. Access-token issuance sits behind an `IAccessTokenService` abstraction so the application layer stays free of the JWT libraries, while the concrete implementation lives at the presentation edge alongside the bearer-validation pipeline.
 
 **Login is split around player selection.** `Login` verifies credentials and returns the issued (pre-selection) token pair plus a summary of each character on the account — it binds no player. A second authenticated `SelectPlayer(playerId, refreshToken)` call then validates that the player belongs to the account (anti-cheat), loads it, binds the session, and **rotates the token pair** to carry the chosen player as a `player_id` claim (consuming the caller's refresh token only after the validations pass, so a rejected selection leaves it intact). The single-active-session takeover check (`ActiveSession`) is a per-player presence check, so it runs *after* selection. This is the seam for multiple characters per account (the in-game character switcher and lossless switching land on top of it).
