+++ docs/backend-sockets.md
@@ -8,7 +8,7 @@ The deep mechanics of the socket layer. The architectural goal — why every in-
 
 ## Single active connection (session takeover warning)
 
-A player may only be connected to one instance at a time: the socket manager records the player's current socket id in Redis and, when a new socket replaces an existing one, signals the old socket to close. To let the client warn the user **before** that takeover happens, the authenticated `Login/ActiveSession` endpoint reports whether the player already has a live socket — deliberately over **HTTP, not a socket command**, because opening a socket to ask the question would itself trigger the replacement it is meant to warn about. Because the check is presence-based it covers "another tab" and "another device" uniformly.
+A player may only be connected to one instance at a time: the socket manager records the player's current socket id in Redis and, when a new socket replaces an existing one, signals the old socket to close. To let the client warn the user **before** that takeover happens, the authenticated `Auth/ActiveSession` endpoint reports whether the player already has a live socket — deliberately over **HTTP, not a socket command**, because opening a socket to ask the question would itself trigger the replacement it is meant to warn about. Because the check is presence-based it covers "another tab" and "another device" uniformly.
 
 The presence key carries a short TTL refreshed on every inbound socket message (the client heartbeats every 10s), so a live connection stays fresh while a connection that vanished without a clean close expires within the TTL. The key is claimed atomically (set-with-TTL in one operation) and rolled back if the post-claim setup fails, so a partial registration can't leave a ghost session that blocks the player.
