+++ docs/frontend-overlays.md
@@ -18,7 +18,7 @@ User-facing notifications (type-aware: error, success, warning, info) are a sing
 Blocking confirmation dialogs are the toast system's counterpart, from the same standardized design. The `modal` store is the single entry point; a single `ModalHost` in the root layout renders the active dialog, and the presentation-only `Modal` flexes across `confirm`/`acknowledge`/`destructive` kinds.
 
 - **Promise-based, await-able API.** `showModal(options)` (and the `confirmModal`/`acknowledgeModal`/`dangerModal` helpers) returns a `Promise<boolean>`, so a caller writes `if (await confirmModal({ … })) { … }`. Used by the game's quit control and the `SocketReplaced` handler (which, when the player's one allowed connection is taken over, stops the engines and acknowledges before returning to login via a client-side navigation — not a reload, which would bounce a still-authenticated client back through the loading screen).
-- **Two sides of the single-connection model.** The `SocketReplaced` handler is the *losing* session's experience; the *winning* side is warned **before** taking over — on login, `confirmSessionTakeover()` calls the authenticated `Login/ActiveSession` endpoint (over HTTP, so it never opens a socket of its own) and confirms if the player already has a live connection elsewhere. It fails open (a non-200 proceeds rather than blocking a legitimate login) and runs only on explicit login, not the silent resume path.
+- **Two sides of the single-connection model.** The `SocketReplaced` handler is the *losing* session's experience; the *winning* side is warned **before** taking over — on login, `confirmSessionTakeover()` calls the authenticated `Auth/ActiveSession` endpoint (over HTTP, so it never opens a socket of its own) and confirms if the player already has a live connection elsewhere. It fails open (a non-200 proceeds rather than blocking a legitimate login) and runs only on explicit login, not the silent resume path.
 - **A queue, not a single slot** — only the head is shown, but requests queue (each with its own resolve) so two callers each get an answer.
 - **Dismissal semantics depend on kind** — backdrop-click/Escape are a cancel for `confirm`/`destructive` but an acknowledgement for `acknowledge`.
 - **The host owns the interaction model so the card stays presentational** — backdrop/Escape dismissal, a focus trap, focus capture+restore, and a scroll lock; it focuses the safe action for a destructive dialog (so a stray Enter can't confirm) and the primary action otherwise. All motion is CSS so `prefers-reduced-motion` collapses it.
