+++ docs/frontend-sockets.md
@@ -6,7 +6,7 @@ The deep mechanics of the frontend API client layer (`$lib/api`). The architectu
 
 - **`token-store.ts`** is the single source of truth for credentials — it persists the `{ accessToken, refreshToken }` pair in local storage (surviving a refresh) and decodes the access token's `exp` for pre-emptive refresh.
 - **`auth.ts`** owns the refresh lifecycle and **collapses concurrent callers onto a single in-flight refresh** — refresh tokens are single-use on the backend, so parallel refreshes would consume the token twice and log the user out. It uses raw `fetch` to stay out of the 401-retry path and avoid an import cycle. That collapsing is per-tab only, while the token pair lives in shared `localStorage` and two tabs are a supported state, so a failed refresh re-reads storage before clearing: if another tab already rotated the token out from under this one, the stored pair is adopted instead of wiping out a session that's still alive elsewhere. That guard only covers `refreshTokens`'s own clear — `handleAuthFailure` (called by the request/socket layers once a refresh is deemed dead) always clears unconditionally, so each caller re-reads storage itself immediately beforehand and skips calling it at all if a concurrent tab has since rotated in a fresh pair.
-- **A refresh failure is only fatal when it is a definitive rejection.** `refreshTokens`/`ensureValidAccessToken` return a discriminated outcome: `rejected` means the backend affirmatively refused the presented refresh token (`Login/Refresh` returns **400** for "Invalid or expired refresh token" — its only failure path; it never 401/403s) or no refresh token is stored at all. Everything else — a network throw, a non-400 non-2xx (a 5xx, the auth rate limiter's 429), or a malformed/unparseable 2xx body — is `retryable`: the stored pair may still be good, so tokens are left in place and no logout is forced. Callers route to `handleAuthFailure` only on `rejected`; on `retryable` the HTTP path surfaces the original 401 to the caller, and the socket path bails out of the connect and lets the keepalive ping's next `ensureSocket()` retry it.
+- **A refresh failure is only fatal when it is a definitive rejection.** `refreshTokens`/`ensureValidAccessToken` return a discriminated outcome: `rejected` means the backend affirmatively refused the presented refresh token (`Auth/Refresh` returns **400** for "Invalid or expired refresh token" — its only failure path; it never 401/403s) or no refresh token is stored at all. Everything else — a network throw, a non-400 non-2xx (a 5xx, the auth rate limiter's 429), or a malformed/unparseable 2xx body — is `retryable`: the stored pair may still be good, so tokens are left in place and no logout is forced. Callers route to `handleAuthFailure` only on `rejected`; on `retryable` the HTTP path surfaces the original 401 to the caller, and the socket path bails out of the connect and lets the keepalive ping's next `ensureSocket()` retry it.
 - **`api-request.ts`** attaches the bearer header, refreshes pre-emptively, and on a 401 refreshes once and retries. The anonymous auth endpoints are exempt.
 - **`api-socket.ts`** passes the access token as a query param (browsers can't set WebSocket handshake headers), refreshes pre-emptively before opening the handshake (mirroring the HTTP path, so a reconnect on an expired token doesn't eat a rejected handshake), and on an auth-rejected handshake refreshes and reconnects (bounded — see below). Because the open path is async, concurrent opens are collapsed onto a single in-flight connect.
